Split test into separate ones.
